kubeadm: refactor upgrade to reduce duplication code
