+++ cmd/kubeadm/app/cmd/phases/upgrade/node/data.go
@@ -18,26 +18,14 @@ limitations under the License.
 package node
 
 import (
-	"io"
-
-	"k8s.io/apimachinery/pkg/util/sets"
-	clientset "k8s.io/client-go/kubernetes"
-
-	kubeadmapi "k8s.io/kubernetes/cmd/kubeadm/app/apis/kubeadm"
+	"k8s.io/kubernetes/cmd/kubeadm/app/cmd/phases/upgrade"
 )
 
 // Data is the interface to use for kubeadm upgrade node phases.
 // The "nodeData" type from "cmd/upgrade/node.go" must satisfy this interface.
 type Data interface {
-	EtcdUpgrade() bool
-	RenewCerts() bool
-	DryRun() bool
-	Cfg() *kubeadmapi.UpgradeConfiguration
-	InitCfg() *kubeadmapi.InitConfiguration
+	upgrade.Data
+
 	IsControlPlaneNode() bool
-	Client() clientset.Interface
-	IgnorePreflightErrors() sets.Set[string]
-	PatchesDir() string
 	KubeConfigPath() string
-	OutputWriter() io.Writer
 }
